rename writer to client
