+++ apps/researcher/src/lib/enricher-instances.ts
@@ -1,7 +1,7 @@
 import {
   EnrichmentFetcher,
   EnrichmentStorer,
-  NanopubWriter,
+  NanopubClient,
 } from '@colonial-collections/enricher';
 import {env} from 'node:process';
 
@@ -9,9 +9,9 @@ export const fetcher = new EnrichmentFetcher({
   endpointUrl: env.NANOPUB_SPARQL_ENDPOINT_URL as string,
 });
 
-const nanopubWriter = new NanopubWriter({
+const nanopubClient = new NanopubClient({
   endpointUrl: env.NANOPUB_WRITE_ENDPOINT_URL as string,
   proxyEndpointUrl: env.NANOPUB_WRITE_PROXY_ENDPOINT_URL as string,
 });
 
-export const storer = new EnrichmentStorer({nanopubWriter});
+export const storer = new EnrichmentStorer({nanopubClient});
