+++ packages/enricher/src/index.ts
@@ -1,3 +1,3 @@
+export * from './client';
 export * from './fetcher';
 export * from './storer';
-export * from './writer';
