+++ packages/enricher/src/storer.integration.test.ts
@@ -1,15 +1,15 @@
 import {AdditionalType} from './definitions';
-import {NanopubWriter} from './writer';
+import {NanopubClient} from './client';
 import {EnrichmentStorer} from './storer';
 import {describe, expect, it} from '@jest/globals';
 import {env} from 'node:process';
 
-const nanopubWriter = new NanopubWriter({
+const nanopubClient = new NanopubClient({
   endpointUrl: env.NANOPUB_WRITE_ENDPOINT_URL as string,
   proxyEndpointUrl: env.NANOPUB_WRITE_PROXY_ENDPOINT_URL as string,
 });
 
-const storer = new EnrichmentStorer({nanopubWriter});
+const storer = new EnrichmentStorer({nanopubClient});
 
 describe('add', () => {
   it('adds a textual enrichment', async () => {
